Rename and change controllers/contacts
